scripts/requirements.txt
# google auth
google-auth-httplib2
google-auth-oauthlib

# google api
google-api-python-client

# development
termcolor
